Add option to check if imaginary modes exist in phonon or QHA calculations
If a user wants to calculate thermochemistry, it must be done on a minimum energy structure otherwise it is not valid. Currently, there is no mechanism for the user to confirm the results are sensible. I have added a keyword argument `imag_mode_threshold: float | None = None` that will raise a `ValueError` if there are imaginary modes with magnitude greater than `imag_mode_threshold`.

Signed-off-by: Andrew S. Rosen <asrosen93@gmail.com>